Use remapped channels in CONDA_CHANNELS
